Only ship the required libraries in the gem artifact
This skips the dev and test deps in the gem artifact and ships a more
mimimal set. This is pretty common in ruby gems and it's something that
Amazon and Microsoft do for their cloud API gems. This drops the
compressed gem size from 56k to 24k and the on disk size from 1.2 megs
to 483k.

Signed-off-by: Tim Smith <tsmith@chef.io>