Drop ActiveSupport as a dependency
Looking through the history of the respository, it seems that keeping
up-to-date with releases of Rails and ActiveSupport is a burden because
there's a lack of trust that new versions of ActiveSupport won't break
compatibility.

Since the usage of ActiveSupport is so minimal in the library and the
test suite, it seems like it would be better to drop it as a dependency
so you don't have that burden anymore.

This has the happy effect that you no longer have to worry about Rails
compatibility and you aren't bringing along a giant library of tools
just for a simple API gem.