Downgrade required bundler version
Downgrades the required version of bundler from 2.1.4 to 2.1.2.

It seems that the Ruby environment that CI runs in cannot access the
latest release of bundler. With that said, 2.1.2 is still very recent
(it was released in December 2019), so I believe that to be an
acceptable version of bundler to use for the sake of running the job to
release the gem on RubyGems.

If this isn't okay, I'll reach out internally to see what we can do
about getting access to the latest release of bundler in the CI
environment.