Allow any Faraday version greater than 0.15
Updates the gemspec to allow any Faraday version greater than 0.15.

Previously, we used the pessimistic operator to only allow versions
greater than or equal to 0.15 and less than 1.0. This changes that to
allow any version greater than 0.15.

I've run the tests with Faraday 1.0.1, and things seems to be okay, but
we'll see what CI says.

---

I also updated our team name, and fixed the spacing of "Digital Ocean" to
"DigitalOcean".